backend/requirements.txt
# Backend requirements
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
gunicorn>=21.2
whitenoise>=6.5.0

# ML / Data
pandas>=2.1
numpy>=1.24
scikit-learn>=1.3
scipy>=1.11

# Utilities
python-dotenv>=1.0
